Moving LocalScope and FramePointer closer to compiler
