+++ engine/runtime/src/main/java/org/enso/compiler/context/FramePointer.java
@@ -1,0 +1,7 @@
+package org.enso.compiler.context;
+
+/**
+ * A representation of a pointer into a stack frame at a given number of levels above the current.
+ */
+public record FramePointer(int parentLevel, int frameSlotIdx) {
+}
