+++ engine/runtime/src/main/java/org/enso/interpreter/node/ClosureRootNode.java
@@ -5,8 +5,8 @@ import com.oracle.truffle.api.dsl.ReportPolymorphism;
 import com.oracle.truffle.api.frame.VirtualFrame;
 import com.oracle.truffle.api.nodes.NodeInfo;
 import com.oracle.truffle.api.source.SourceSection;
+import org.enso.compiler.context.LocalScope;
 import org.enso.interpreter.EnsoLanguage;
-import org.enso.interpreter.runtime.scope.LocalScope;
 import org.enso.interpreter.runtime.scope.ModuleScope;
 
 /**
