+++ engine/runtime/src/main/scala/org/enso/compiler/context/InlineContext.scala
@@ -1,9 +1,9 @@
 package org.enso.compiler.context
 
 import org.enso.compiler.PackageRepository
+import org.enso.compiler.context.LocalScope
 import org.enso.compiler.data.CompilerConfig
 import org.enso.compiler.pass.PassConfiguration
-import org.enso.interpreter.runtime.scope.LocalScope
 
 /** A type containing the information about the execution context for an inline
   * expression.
